Home page basic completed. Footer content added
Home page basic content's functionality has been implemented. Need to add the interaction of clicking from a search result in the search bar and directed to a corresponding place page. Need to add the interaction of clicking on a title of a recommendation place and is directed to a corresponding place page.